updates for v8.8.9
- update Cake to 0.13.24 (adds C99 VLA support).
- bcc now sorts files alphabetically before processing (so that Windows and Linux will process files in the same order).
- minor code generator changes to make Linux and Windows generate identical code.
- recompile C libraries (a precaution due to code generator changes).
- remove Lua 5.4.4 from C libraries
- minor catapult document updates.

+++ source/lcc/src/simp.c
@@ -253,7 +253,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			xcvtcnst(P,l->u.v.p,ty,p,l->u.v.p);
 			break;
 		case CVI+U:
-			xcvtcnst(I,l->u.v.i,ty,u,((unsigned long)l->u.v.i)&ones(8*ty->size));
+			xcvtcnst(I,(unsigned)l->u.v.i,ty,u,((unsigned long)l->u.v.i)&ones(8*ty->size));
 			break;
 		case CVU+U:
 			xcvtcnst(U,l->u.v.u,ty,u,l->u.v.u&ones(8*ty->size));
